move server_logs to handlers, update channel deletion methods, create SERVER_LOGS_CHANNEL constant
