+++ main.py
@@ -35,7 +35,6 @@ from commands.reports import reports
 from commands.quiz import quiz
 from commands.report import report
 from commands.scores import scores
-from commands.server_logs import show_leave_message, show_nick_change_message
 from commands.setwager import setwager
 from commands.shop import shop
 from commands.slots import slots, testslots
@@ -45,6 +44,7 @@ from commands.update_status import update_status
 # Handlers
 from handlers.alerts import handle_alerts
 from handlers.bot_autoresponse import handle_bot_affirmations
+from handlers.server_logs import *
 from handlers.starboard import get_all_starboard_posts, handle_starboard_reactions
 from handlers.xp import handle_message_xp, handle_react_xp
 
@@ -196,19 +196,15 @@ async def on_member_update(memberBefore,memberAfter):
 # Listen to channel updates
 @bot.event
 async def on_guild_channel_create(channel):
-  dev_channel = await bot.fetch_channel(DEV_CHANNEL)
-  await dev_channel.send(f"✨ A new channel, **'#{channel.name}'**, was just created!")
+  await show_channel_creation_message(channel)
 
 @bot.event
 async def on_guild_channel_delete(channel):
-  dev_channel = await bot.fetch_channel(DEV_CHANNEL)
-  await dev_channel.send(f"💥 **'#{channel.name}'**, was just deleted!")
+  await show_channel_deletion_message(channel)
 
 @bot.event
 async def on_guild_channel_update(before, after):
-  dev_channel = await bot.fetch_channel(DEV_CHANNEL)
-  if before.name != after.name:
-    await dev_channel.send(f"↪️ **'#{before.name}'** was changed to **'#{after.name}'**!")
+ await show_channel_rename_message(before, after)
 
 # listen to interaction errors
 @bot.event
